+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v13.1.0
 
+- Feat: add new `StateManager.Init` to init mutations using `UnityAction`.
 - Fix: `ICommand` no longer inherits `IEvent`. [#44](https://github.com/DiscreteTom/UniStart/issues/44)
 - Fix: `ICommandBus.Push` will throw exception if the command is not registered. [#45](https://github.com/DiscreteTom/UniStart/issues/45)